+++ app/Http/Controllers/HomeController.php
@@ -8,6 +8,7 @@ use App\Models\User;
 use App\Models\HelpdeskTicket;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
 class HomeController extends Controller
@@ -134,6 +135,15 @@ class HomeController extends Controller
 
     private function regularDashboard()
     {
+        if (
+            Auth::user()->type === 'client'
+            && Auth::user()->hasRole('production-client')
+            && Auth::user()->can('view-video-production-dashboard')
+            && Auth::user()->can('view-video-production')
+        ) {
+            return Inertia::render('dashboard');
+        }
+
         $packagesPath = base_path('packages/workdo');
 
         // find dashboard menu from all  active package and redirect if found